envs/cwah/requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.6.2
anytree==2.8.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.11.2
bleach==6.0.0
braceexpand==0.1.7
build==0.10.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dgl==0.9.1
distlib==0.3.6
dnspython==2.3.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.0
executing==1.2.0
ExifRead-nocycle==3.0.1
fastjsonschema==2.16.2
filelock==3.8.2
fire==0.4.0
Flask==2.2.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.51.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.7.0
huggingface-hub==0.11.1
idna==2.8
imageio==2.25.0
img2dataset==1.41.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflect==6.0.2
ipdb==0.13.11
ipykernel==6.21.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.0
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.1.3
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kernel-driver==0.0.7
kiwisolver==1.4.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mkl-fft
mkl-random
mkl-service==2.4.0
mmcv==2.0.0
mmengine==0.7.2
model-index==0.1.11
msgpack==1.0.4
multidict==6.0.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.5.0
nbterm==0.0.13
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy
oauthlib==3.2.2
open3d==0.16.0
openai==0.27.0
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
openmim==0.3.7
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip-tools==6.13.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plotly==5.13.1
prettytable==3.6.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pudb==2022.1.3
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pydantic==1.10.5
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyquaternion==0.9.9
pyrsistent==0.19.3
python-json-logger==2.0.5
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
ray==2.2.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.2.0
rouge-score==0.1.2
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.14.0
setproctitle==1.3.2
shapely==2.0.1
shortuuid==1.0.11
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.2.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.1.23.1
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.12.1
tomli==2.0.1
torch
torchaudio
torchvision
tornado==6.2
tqdm
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.15
urwid==2.1.2
urwid-readline==0.13
virtualenv==20.17.1
wandb==0.12.21
wcwidth==0.2.5
webcolors==1.12
webdataset==0.2.31
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
yacs==0.1.8
yapf==0.32.0
yarg==0.1.9
yarl==1.5.1
zipp==3.3.0
pyautogen
transformers
accelerate>=0.20.3
